pyproject.toml
[project]
name = "mini-wikipedia-rag"
version = "0.1.0"
description = "Example RAG environment using a few wikipedia articles with Llama Index"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "datasets>=3.0.0",
    "fastapi>=0.121.3",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "jupyterlab>=4.5.6",
    "llama-index-core>=0.14.0",
    "llama-index-embeddings-azure-openai>=0.4.1",
    "llama-index-llms-azure-openai>=0.4.2",
    "numpy>=2.3.5",
    "pandas>=2.0.0",
    "uvicorn>=0.30.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "pandas>=3.0.1",
    "numpy>=2.3.5",
    "jupyter>=1.1.1",
    "ipykernel>=7.2.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
# Filter out known warnings from dependencies
filterwarnings = [
    "ignore::DeprecationWarning:llama_index.*",
    "ignore:There is no current event loop:DeprecationWarning"
]
